Fix some MSVC warnings with MSVC 17.14.16 in the 32-bit i386 build.
src\rp-download\CurlDownloader.cpp(255,16): warning C4244: 'argument': conversion from 'const off64_t' to 'unsigned int', possible loss of data

src\librpbase\RomMetaData.cpp(182,42): warning C4267: '=': conversion from 'size_t' to '_Ty', possible loss of data
src\librpbase\RomMetaData.cpp(182,42): warning C4267:         with
src\librpbase\RomMetaData.cpp(182,42): warning C4267:         [
src\librpbase\RomMetaData.cpp(182,42): warning C4267:             _Ty=int8_t
src\librpbase\RomMetaData.cpp(182,42): warning C4267:         ]

src\libromdata\disc\NCCHReader.cpp(655,50): warning C4244: '=': conversion from 'off64_t' to 'uint32_t', possible loss of data

Also fix a code analyzer warning:

lnt-arithmetic-overflow: A sub-expression may overflow before being assigned to a wider type.
src\libromdata\disc\NCCHReader.cpp(685)

+++ src/libromdata/disc/NCCHReader.cpp
@@ -652,7 +652,7 @@ int NCCHReader::seek(off64_t pos, SeekWhence whence)
 		m_lastError = EINVAL;
 		return -1;
 	}
-	d->pos = static_cast<off64_t>(constrain_file_pos(pos, static_cast<off64_t>(d->ncch_length)));
+	d->pos = static_cast<uint32_t>(constrain_file_pos(pos, static_cast<off64_t>(d->ncch_length)));
 	return 0;
 }
 
@@ -682,7 +682,7 @@ off64_t NCCHReader::size(void)
 {
 	// TODO: Errors?
 	RP_D(const NCCHReader);
-	const off64_t ret = d->ncch_length - sizeof(N3DS_NCCH_Header_t);
+	const off64_t ret = static_cast<off64_t>(d->ncch_length) - sizeof(N3DS_NCCH_Header_t);
 	return (ret >= 0 ? ret : 0);
 }
